refactor: simplify frontend
